+++ test/trunc.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2008-2011, 2013 D. V. Wiebe
+/* Copyright (C) 2008-2011, 2013, 2015 D. V. Wiebe
  *
  ***************************************************************************
  *
@@ -57,7 +57,8 @@ int main(void)
 
   stat_format = stat(format, &buf);
   CHECKI(stat_format, 0);
-  CHECK((buf.st_size > 0),buf.st_size,"%lli","%s",(long long)buf.st_size,"> 0");
+  CHECK((buf.st_size > 0),buf.st_size,"%" PRId64,"%s",(int64_t)buf.st_size,
+      "> 0");
 
   unlink(format);
   rmdir(filedir);
